Extend model chain to order-11; shrink StateMap count to u8
Cumulative enwik8 sweep atop 33a4aa1 (keep a model only if it lowers
whole-stream bpb): order8/9/10/11 each earn their place and are added
default-on, taking enwik8 1.6893 -> 1.6843 bpb (hutter 1.8177). order12
regressed (+97 B) and sparse24 stayed negative-weight, so both were left
out. Default profile bitmask is now 0x7FFFF (19 features).

StateMap's per-state observation count is capped at LIMIT=11, so store it
in a Vec<u8> instead of Vec<u16>: value-identical arithmetic (bpb
unchanged, roundtrip preserved), saving 1 byte/slot -- ~256 MiB per hashed
table on enwik9.

+++ src/models/statemap.rs
@@ -36,7 +36,7 @@ fn rate_table() -> &'static [i32; LIMIT + 1] {
 #[derive(Debug)]
 pub(crate) struct StateMap {
     p: Vec<i32>, // 16-bit probability per state
-    n: Vec<u16>, // observation count per state (capped at LIMIT)
+    n: Vec<u8>,  // observation count per state (capped at LIMIT, which fits a byte)
 }
 
 impl StateMap {
